private/requirements_web.txt
# Forest Fire Emergency Web Interface Requirements
flask==3.0.0
werkzeug==3.0.1
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# AI Integration
google-generativeai==0.3.2

# Database
sqlite3

# Development
python-dotenv==1.0.0
